+++ www/go/modules/community/addressbook/install/install.sql
@@ -98,7 +98,9 @@ CREATE TABLE `addressbook_url` (
 
 CREATE TABLE `addressbook_user_settings` (
   `userId` int(11) NOT NULL,
-  `defaultAddressBookId` int(11) DEFAULT NULL
+  `defaultAddressBookId` int(11) DEFAULT NULL,
+  `salutationTemplate` TEXT NOT NULL, 
+  `sortBy` ENUM('name','lastName') NOT NULL DEFAULT 'name'
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci ROW_FORMAT=COMPACT;