+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Address book: Added organization filter for contacts "org"
+
 01-07-2019 6.4.25
 - Core: Simplified search. All words will be used with AND instead of OR. A wildcard will be placed after each word. 
 - Core: Fixed upgrade error can't find module core/groups